refactor: refactoring route-based role
