+++ src/pages/DashboardPage.tsx
@@ -24,7 +24,7 @@ import { useNavigate } from "react-router";
 
 export default function DashboardPage() {
   // set role = driver | warehouse | outlet
-  const [role, useRole] = useState("driver");
+  const [role, useRole] = useState("warehouse");
   // set username driver kalo role === driver
   const [usernameDriver, setUsernameDriver] = useState("budi_cepat");
