+++ src/pages/warehouse/OutletPage.tsx
